refactor: review comments, docs
- address review comments
- update docs

Contributes to: #13
