+++ client/Hooks/index.ts
@@ -1,0 +1,5 @@
+/*
+ * Copyright Strimzi authors.
+ * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
+ */
+export * from './useConfigFeatureFlag/useConfigFeatureFlag';
