+++ client/Hooks/useConfigFeatureFlag/README.md
@@ -1,0 +1,3 @@
+# useConfigFeatureFlag
+
+The `useConfigFeatureFlag` hook provides access to the value provided by the `ConfigFeatureFlag` context. [Further details about the value returned can be found here](../../Contexts/ConfigFeatureFlag/README.md).
