+++ client/Panels/Home/Home.tsx
@@ -3,20 +3,22 @@
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
 import React, { FunctionComponent } from 'react';
+import get from 'lodash.get';
 import image from 'Images/logo.png';
 import './style.scss';
-import { useConfigFeatureFlag, FeatureFlag } from 'Contexts';
+import { useConfigFeatureFlag } from 'Hooks';
 
 const Home: FunctionComponent = ({ children }) => {
-  const { client, isComplete } = useConfigFeatureFlag();
+  const { client, featureFlags, isComplete } = useConfigFeatureFlag();
+  const version = get(client, 'about.version', '');
+  // use the feature flag from context - could also use the `FeatureFlag` component - this just shows alternative usage
+  const showVersion = get(featureFlags, 'client.Home.showVersion', false);
 
   return (
     <div className='home'>
       <img src={image} alt='Strimzi logo' />
       Welcome to the Strimzi UI
-      <FeatureFlag flag={'client.Home.showVersion'}>
-        {isComplete && `Version: ${client.version}`}
-      </FeatureFlag>
+      {showVersion && isComplete && `Version: ${version}`}
       {children}
     </div>
   );
