+++ client/Queries/Config/index.ts
@@ -7,12 +7,17 @@ import gql from 'graphql-tag';
 export const GET_CONFIG = gql`
   query {
     client {
-      version
+      about {
+        version
+      }
     }
     featureFlags {
       client {
         Home {
           showVersion
+        }
+        Pages {
+          PlaceholderHome
         }
       }
     }
