+++ package.json
@@ -86,6 +86,7 @@
     "@storybook/react": "^6.0.28",
     "@testing-library/jest-dom": "^5.11.5",
     "@testing-library/react": "^11.1.1",
+    "@testing-library/react-hooks": "^3.4.2",
     "@testing-library/user-event": "^12.2.2",
     "@types/cypress-cucumber-preprocessor": "^4.0.0",
     "@types/express": "^4.17.8",
@@ -120,6 +121,7 @@
     "nodemon": "^2.0.6",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.1.2",
+    "react-test-renderer": "^16.14.0",
     "sass-loader": "^10.0.5",
     "storybook": "^6.0.28",
     "storybook-readme": "^5.0.8",
